// frontend/src/app/api/model/moduleMemberDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ModuleMemberDto { 
    userId?: string;
    fullName?: string | null;
    studentId?: string | null;
    email?: string | null;
}